refactor div in component the task
