+++ src/app/components/list-tasks/list-tasks.component.html
@@ -1,20 +1,7 @@
 <section class="main">
   <ul class="todo-list">
     @for (task of filteredTasks(); track $index) {
-      <li 
-      [class.completed]="task.completed"
-      [class.editing]="task.editing"
-      (dblclick)="editingTask($index, $event)">
-        <div class="view">
-          <input class="toggle" type="checkbox" [checked]="task.completed"
-          (click)="handleCompletedTask($index)"/>
-          <label>{{task.title}}</label>
-          <button class="destroy"></button>
-        </div>
-        <input class="edit" [value]="task.title"
-        (keydown.enter)="updateTextTask($event, $index)"
-        (keydown.esc)="exitEditingTask()"/>
-      </li>
+      <app-task [task]="task" [index]="$index"></app-task>
     }
   </ul>
 </section>
