+++ src/app/components/list-tasks/list-tasks.component.ts
@@ -1,10 +1,11 @@
 import { Component, inject } from '@angular/core';
 import { TasksService } from 'src/app/services/tasks.service';
+import { TaskComponent } from '../task/task.component';
 
 @Component({
   selector: 'app-list-tasks',
   standalone: true,
-  imports: [],
+  imports: [TaskComponent],
   templateUrl: './list-tasks.component.html',
   styleUrl: './list-tasks.component.css'
 })
@@ -16,42 +17,6 @@ export class ListTasksComponent {
 
   handleCompletedTask(index: number){
     this.taskService.changeCompletedTask(index);
-  }
-
-  editingTask(index: number, event: Event){
-    this.tasks.update((tasks) => {
-      return tasks.map((task, position) => {
-        if (position === index){
-          return {
-            ...task,
-            editing: true
-          }
-        }
-        return {
-          ...task,
-          editing: false
-        };
-      })
-    })
-  }
-  
-  exitEditingTask(){
-    this.tasks.update((tasks) => {
-      return tasks.map((task) => {
-        return {
-          ...task,
-          editing: false
-        };
-      })
-    })
-  }
-
-  updateTextTask(event: Event, index: number){
-    const newTask = (event.target as HTMLInputElement).value.trim();
-    
-    if(newTask !== ''){
-      this.taskService.updateTask(newTask, index);
-    }
   }
  
 }
