+++ src/app/components/task/task.component.html
@@ -1,0 +1,14 @@
+<li 
+    [class.completed]="task.completed"
+    [class.editing]="task.editing"
+    (dblclick)="handleEditingTask()">
+    <div class="view">
+        <input class="toggle" type="checkbox" [checked]="task.completed"
+        (click)="handleCompletedTask()"/>
+        <label>{{task.title}}</label>
+        <button class="destroy"></button>
+    </div>
+    <input class="edit" [value]="task.title"
+    (keydown.enter)="handleUpdateTextTask($event)"
+    (keydown.esc)="handleExitEditingTask()"/>
+</li>
